\name{Acari}
\alias{Acari}
\title{Acari Related Metrics

Calculates various metrics based on Acari taxa}
\usage{
  Acari_Percent(x) Acari_PercentTaxa(x) Acari_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Acari Related Metrics

  Calculates various metrics based on Acari taxa
}

