% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate.BMI}
\alias{aggregate.BMI}
\title{Aggregate FinalID}
\usage{
\method{aggregate}{BMI}(x, effortlevel = c("SAFIT1", "SAFIT2"))
}
\arguments{
\item{x}{An object of class BMI}
}
\description{
Aggregates to  taxonomic effort levels SAFIT1 and SAFIT2
}

