% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BMICSCI}
\alias{BMICSCI}
\title{Compute BMI metrics for CSCI

Computes all CSCI related BMI metrics. This includes, clinger percent taxa, coleoptera,
percent taxa, taxnomic richness, EPT percent taxa, number of shredder taxa, and percent
of individuals that are intolerant.}
\usage{
BMICSCI(x, effort = 2)
}
\arguments{
\item{x}{An object of class BMIagg or BMIprc}
}
\description{
Compute BMI metrics for CSCI

Computes all CSCI related BMI metrics. This includes, clinger percent taxa, coleoptera,
percent taxa, taxnomic richness, EPT percent taxa, number of shredder taxa, and percent
of individuals that are intolerant.
}

