\name{Burrower}
\alias{Burrower}
\title{Burrower

Calculates metrics related to burrowing.}
\usage{
  Burrower_Percent(x) Burrower_PercentTaxa(x)
    Burrower_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Burrower

  Calculates metrics related to burrowing.
}

