\name{CFCG}
\alias{CFCG}
\title{CFCG

Calculates metrics related to collector foragers/ collector gatherers.}
\usage{
  CFCG_Percent(x) CFCG_PercentTaxa(x) CFCG_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  CFCG

  Calculates metrics related to collector foragers/
  collector gatherers.
}

