\name{Chironomidae}
\alias{Chironomidae}
\title{Chironomidae related metrics}
\usage{
  Chironomidae_PercentTaxa(x) Chironomidae_Taxa(x)
    Chironominae_Percent(x) Chironominae_PercentOfMidges(x)
    Tanypodinae_Percent(x) Tanypodinae_PercentOfMidges(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Calculates metrics related to chironomids.
}

