\name{Climber}
\alias{Climber}
\title{Climber

Calculates metrics related to the climber habit group.}
\usage{
  Climber_Percent(x) Climber_PercentTaxa(x) Climber_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Climber

  Calculates metrics related to the climber habit group.
}

