\name{Clinger}
\alias{Clinger}
\title{Clinger

Calculates metrics related to the clinger habit group.}
\usage{
  Clinger_Percent(x) Clinger_PercentTaxa(x) Clinger_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Clinger

  Calculates metrics related to the clinger habit group.
}

