\name{Coleoptera_Percent}
\alias{Coleoptera_Percent}
\title{Coleoptera Related Metrics}
\usage{
  Coleoptera_Percent(x) Coleoptera_PercentTaxa(x)
    Coleoptera_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Calculates various metrics based on Coleoptera taxa
}

