\name{Crustacea_Percent}
\alias{Crustacea_Percent}
\title{Crustacea Related Metrics}
\usage{
  Crustacea_Percent(x, ...)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}

  \item{level}{Either "SAFIT1" or "SAFIT2"}
}
\description{
  Calculates various metrics based on Crustacea taxa
}

