\name{Diptera_Percent}
\alias{Diptera_Percent}
\title{Diptera Related Metrics}
\usage{
  Diptera_Percent(x) Diptera_PercentTaxa(x) Diptera_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Calculates various metrics based on Diptera taxa
}

