\name{Ephemeroptera}
\alias{Ephemeroptera}
\title{Ephemeroptera Related Metrics

Calculates various metrics based on Ephemeroptera taxa}
\usage{
  Ephemeroptera_Percent(x) Ephemeroptera_PercentTaxa(x)
    Ephemeroptera_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Ephemeroptera Related Metrics

  Calculates various metrics based on Ephemeroptera taxa
}

