\name{EPT}
\alias{EPT}
\title{EPT Related Metrics

Calculates various metrics based on EPT taxa}
\usage{
  EPT_Percent(x) EPT_PercentTaxa(x) EPT_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  EPT Related Metrics

  Calculates various metrics based on EPT taxa
}

