\name{Noninsect}
\alias{Noninsect}
\title{Noninsect Related Metrics

Calculates various metrics based on Noninsect taxa}
\usage{
  Noninsect_Percent(x) Noninsect_PercentTaxa(x)
    Noninsect_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Noninsect Related Metrics

  Calculates various metrics based on Noninsect taxa
}

