\name{Plecoptera}
\alias{Plecoptera}
\title{Plecoptera Related Metrics

Calculates various metrics based on Plecoptera taxa}
\usage{
  Plecoptera_Percent(x) Plecoptera_PercentTaxa(x)
    Plecoptera_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Plecoptera Related Metrics

  Calculates various metrics based on Plecoptera taxa
}

