\name{Scraper}
\alias{Scraper}
\title{Scraper related metrics

Calculates metrics related to the Scraper feeding group.}
\usage{
  Scraper_Percent(x) Scraper_PercentTaxa(x) Scraper_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Scraper related metrics

  Calculates metrics related to the Scraper feeding group.
}

