\name{Shredder}
\alias{Shredder}
\title{Shredder related metrics

Calculates metrics related to the Shredder feeding group.}
\usage{
  Shredder_Percent(x) Shredder_PercentTaxa(x)
    Shredder_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Shredder related metrics

  Calculates metrics related to the Shredder feeding group.
}

