\name{Swimmer}
\alias{Swimmer}
\title{Swimmer related metrics

Calculates metrics related to the Swimmer habit group.}
\usage{
  Swimmer_Percent(x) Swimmer_PercentTaxa(x) Swimmer_Taxa(x)
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Swimmer related metrics

  Calculates metrics related to the Swimmer habit group.
}

