\name{Tolerance}
\alias{Intolerant}
\alias{Tolerance}
\alias{Tolerant}
\title{Tolerance related metrics

Calculates metrics related to tolerance values.}
\usage{
  Intolerant_Percent(x) Intolerant_PercentTaxa(x)
    Intolerant_Taxa(x) Tolerant_Percent(x)
    Tolerant_PercentTaxa(x) Tolerant_Taxa(x)
}
\description{
  Tolerance related metrics

  Calculates metrics related to tolerance values.
}

