\name{Total_Taxa}
\alias{Total_Taxa}
\title{Benthic Community Metrics

Calculates metrics concerning total number of taxa and sample diversity.}
\usage{
  Total_Taxa Shannon_Diversity Simpson_Diversity
}
\arguments{
  \item{x}{An object of class BMIprc or BMIagg}

  \item{level}{Either "SAFIT1" or "SAFIT2"}
}
\description{
  Benthic Community Metrics

  Calculates metrics concerning total number of taxa and
  sample diversity.
}

