\name{Trichoptera}
\alias{Trichoptera}
\title{Trichoptera Related Metrics

Calculates various metrics based on Trichoptera taxa}
\usage{
  Trichoptera_Percent(x) Trichoptera_PercentTaxa(x)
    Trichoptera_Taxa(x)
}
\arguments{
  \item{level}{Either "SAFIT1" or "SAFIT2"}

  \item{x}{An object of class BMIprc or BMIagg}
}
\description{
  Trichoptera Related Metrics

  Calculates various metrics based on Trichoptera taxa
}

